2.1-databases/work_with_database/requirements.txt
django~=3.2.8
psycopg2-binary
awesome-slugify

pip~=21.3.1
setuptools~=49.2.1